Refactor test functions
