+++ xobjects/sparse/__init__.py
@@ -1,3 +1,3 @@
-from ._sparse import factorized_sparse_solver
+from ._sparse import factorized_sparse_solver, rel_residual
 from . import solvers
-__all__ = ["factorized_sparse_solver","solvers"]
+__all__ = ["factorized_sparse_solver","solvers", "rel_residual"]
